Starting to work on force

+++ Vector2D.java
@@ -1,9 +1,9 @@
 public class Vector2D
 {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     
-    public Vector2D(int x1, int y1)
+    public Vector2D(double x1, double y1)
     {
         x = x1;
         y = y1;
